agents/requirements.txt
fastapi==0.115.6
black==24.10.0
langgraph==0.3.0
langchain==0.3.14
langchain-community==0.3.14
langchain-openai==0.3.1
tavily-python==0.5.1
python-dotenv==1.0.1
pymilvus==2.5.4
uvicorn==0.34.0
transformers==4.47.1
accelerate==1.3.0
sentence-transformers==3.3.1
einops